extract document details panel into separate component + ui changes
