+++ app/components/documents/document-preview/details-tab.hbs
@@ -1,101 +1,6 @@
 <div class="auk-u-m-4">
-  <Auk::KeyValue
-    data-test-details-tab-name
-    @key={{t "document-name"}}
-  >
-    {{#if this.isEditingDetails}}
-      <Textarea
-        data-test-details-tab-edit-name
-        aria-labelledby="piece-name"
-        class="auk-textarea"
-        @value={{mut @piece.name}}
-      />
-    {{else}}
-      {{@piece.name}}
-    {{/if}}
-  </Auk::KeyValue>
-  {{#if (or this.documentType this.isEditingDetails)}}
-    <Auk::KeyValue
-      data-test-details-tab-document-type
-      @key={{t "document-type"}}
-    >
-      {{#if this.isEditingDetails}}
-        <Utils::ModelSelector
-          data-test-details-tab-edit-document-type
-          @modelName="document-type"
-          @allowClear={{true}}
-          @searchField="label"
-          @displayField="label"
-          @sortField="priority"
-          @selected={{this.documentType}}
-          @onChange={{this.setDocumentType}}
-        />
-      {{else}}
-        {{this.documentType.label}}
-      {{/if}}
-    </Auk::KeyValue>
-  {{/if}}
-  <Auk::KeyValue 
-    data-test-details-tab-access-level
-    @key={{t "publicity"}}
-  >
-    {{#if this.isEditingDetails}}
-      <Utils::AccessLevelSelector
-        data-test-details-tab-edit-access-level
-        @allowClear={{true}}
-        @displayField="label"
-        @sortField="position"
-        @selected={{this.accessLevel}}
-        @onChange={{this.setAccessLevel}}
-      />
-    {{else}}
-      <Utils::AccessLevelColorBadge @accessLevel={{this.accessLevel}} />
-    {{/if}}
-  </Auk::KeyValue>
-  <Auk::KeyValue @key={{t "file-type"}}>
-    <Utils::Documents::FileTypePill
-      data-test-details-tab-file-type
-      @extension={{@piece.file.extension}}
-    />
-  </Auk::KeyValue>
-  {{#if this.currentSession.isEditor}}
-    <Auk::Toolbar as |toolbar|>
-      <toolbar.Group @position="right" as |group|>
-        {{#if this.isEditingDetails}}
-          <group.Item>
-            <Auk::ButtonLink
-              data-test-details-tab-cancel
-              @disabled={{this.isProcessing}}
-              @loading={{this.cancelEditDetails.isRunning}}
-              {{on "click" (perform this.cancelEditDetails)}}
-            >
-              {{t "cancel"}}
-            </Auk::ButtonLink>
-          </group.Item>
-          <group.Item>
-            <Auk::Button
-              data-test-details-tab-save
-              @skin="primary"
-              @disabled={{this.isProcessing}}
-              @loading={{this.saveDetails.isRunning}}
-              {{on "click" (perform this.saveDetails)}}
-            >
-              {{t "save"}}
-            </Auk::Button>
-          </group.Item>
-        {{else}}
-          <group.Item>
-            <Auk::Button
-              data-test-details-tab-edit
-              @skin="tertiary"
-              @icon="pencil"
-              {{on "click" this.openEditDetails}}
-            >
-              {{t "edit-2"}}
-            </Auk::Button>
-          </group.Item>
-        {{/if}}
-      </toolbar.Group>
-    </Auk::Toolbar>
-  {{/if}}
+  <Documents::DocumentDetailsPanel
+    @piece={{@piece}}
+    @documentContainer={{@documentContainer}}
+  />
 </div>
