+++ app/components/documents/document-preview/document-preview-sidebar.hbs
@@ -2,7 +2,7 @@
   @size="xlarge"
   @position="right"
   @skin="white"
-  class="auk-u-pb-6 auk-o-flex auk-o-flex--vertical
+  class="auk-u-pb-6 auk-o-flex auk-o-flex--vertical auk-u-sidebar-min-width
     {{if @isOpen "auk-scroll-wrapper"}}"
   @isCollapsed={{not @isOpen}}
 >
