+++ app/styles/au/_au-components.scss
@@ -14,6 +14,7 @@
 @import 'ember-appuniversum/c-toolbar';
 @import 'ember-appuniversum/c-input';
 @import 'ember-appuniversum/c-loader';
+@import 'ember-appuniversum/c-modal';
 
 
 // Overrides
