+++ translations/nl-be.json
@@ -1053,5 +1053,8 @@
   "open-search-for-mandatee": "Vrij zoeken",
   "open-search-for-mandatee-tooltip": "Opgelet. Activeer deze optie enkel indien strikt noodzakelijk. In tegenstelling tot bij standaard zoeken, kunnen de aangeboden resultaten bij vrij zoeken ook ministers bevatten die niet actief zijn/waren gedurende de referentieperiode waarvoor u zoekt. Deze functionaliteit kan bv. gebruikt worden om net-niet-meer actieve ministers (rond een ministerwissel) toe te wijzen. ",
   "fullscreen": "Volledig scherm",
-  "close-fullscreen": "Sluit volledig scherm"
+  "close-fullscreen": "Sluit volledig scherm",
+  "document-properties": "Documenteigenschappen",
+  "replace": "Vervangen",
+  "source-file": "Bronbestand"
 }
